@@ -5,24 +5,43 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.4</version>
+        <version>3.2.6</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>uk.nhs.england</groupId>
     <artifactId>iops-fhir-mhd</artifactId>
-    <version>1.0.9</version>
+    <version>7.6.0</version>
     <name>iops-fhir-mhd</name>
     <description>IOPS FHIR MHD</description>
 
     <properties>
-        <java.version>1.8</java.version>
-        <kotlin.version>1.5.10</kotlin.version>
-        <fhir.version>6.2.5</fhir.version>
-        <swagger_version>2.1.12</swagger_version>
-        <parser_version>2.0.30</parser_version>
+        <java.version>20</java.version>
+        <kotlin.version>1.9.22</kotlin.version>
+        <fhir.version>7.6.0</fhir.version>
+        <swagger_version>2.2.25</swagger_version>
+        <parser_version>2.1.20</parser_version>
+        <jackson_databind_version>2.16.1</jackson_databind_version>
+        <jaxb_api_version>2.3.1</jaxb_api_version>
     </properties>
 
     <dependencies>
+        <dependency>
+            <groupId>org.jetbrains.kotlin</groupId>
+            <artifactId>kotlin-test-junit5</artifactId>
+            <version>1.8.0</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter-engine</artifactId>
+            <version>5.8.2</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.jetbrains.kotlin</groupId>
+            <artifactId>kotlin-stdlib</artifactId>
+        </dependency>
+
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
