@@ -12,7 +12,7 @@ import org.springframework.web.cors.CorsConfiguration
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource
 import uk.nhs.england.mhd.interceptor.CognitoAuthInterceptor
 import uk.nhs.england.mhd.util.CorsFilter
-import javax.servlet.Filter
+import jakarta.servlet.Filter
 
 @Configuration
 open class ApplicationConfiguration {
@@ -37,18 +37,6 @@ open class ApplicationConfiguration {
         return client
     }
 
-    @Bean
-    fun corsFilter(): FilterRegistrationBean<*>? {
-        val source = UrlBasedCorsConfigurationSource()
-        val config = CorsConfiguration()
-        config.allowCredentials = true
-        config.addAllowedOrigin("*")
-        config.addAllowedHeader("*")
-        config.addAllowedMethod("*")
-        source.registerCorsConfiguration("/**", config)
-        val bean: FilterRegistrationBean<*> = FilterRegistrationBean<Filter>(CorsFilter())
-        bean.order = 0
-        return bean
-    }
+
 
 }
