@@ -17,9 +17,9 @@ import uk.nhs.england.mhd.configuration.FHIRServerProperties
 import uk.nhs.england.mhd.util.FhirSystems
 import java.io.IOException
 import java.util.*
-import javax.servlet.ServletException
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.servlet.ServletException
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 
 @Interceptor
 class AWSAuditEventLoggingInterceptor(
