@@ -11,7 +11,7 @@ import org.hl7.fhir.r4.model.*
 import org.springframework.stereotype.Component
 import uk.nhs.england.mhd.awsProvider.AWSBinary
 import uk.nhs.england.mhd.interceptor.CognitoAuthInterceptor
-import javax.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletRequest
 
 @Component
 class BinaryProvider(var cognitoAuthInterceptor: CognitoAuthInterceptor, val awsBinary: AWSBinary) : IResourceProvider {
