@@ -10,7 +10,7 @@ import org.hl7.fhir.r4.model.*
 import org.springframework.stereotype.Component
 import uk.nhs.england.mhd.awsProvider.AWSDocumentReference
 import uk.nhs.england.mhd.interceptor.CognitoAuthInterceptor
-import javax.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletRequest
 
 @Component
 class DocumentReferenceProvider(var cognitoAuthInterceptor: CognitoAuthInterceptor,
