@@ -16,12 +16,13 @@ class CapabilityStatementInterceptor(
 
     @Hook(Pointcut.SERVER_CAPABILITY_STATEMENT_GENERATED)
     fun customize(theCapabilityStatement: IBaseConformance) {
-
-        // Cast to the appropriate version
         val cs: CapabilityStatement = theCapabilityStatement as CapabilityStatement
-
+        cs.name = fhirServerProperties.server.name
+        cs.software.name = fhirServerProperties.server.name
+        cs.software.version = fhirServerProperties.server.version
+        cs.publisher = "NHS England"
         cs.implementation.url = fhirServerProperties.server.baseUrl
-        cs.implementation.description = "NHS Digital UKCore API Reference Implementation"
+        cs.implementation.description = fhirServerProperties.server.name
     }
 
     fun getResourceComponent(type : String, cs : CapabilityStatement ) : CapabilityStatement.CapabilityStatementRestResourceComponent? {
